android: Move AndroidOptions out of RedmadrobotExtension
